+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2026-03-01]
+### Added
+- Added `selector_cal_distance` to AFC_lane for units like ViViD that have selectors. If this value is set AFC will move the selector the supplied distance in mm. This is to help make sure selectors have a good grip on the filament.
+### Update
+- Added macro AFC_RESET language to message when failing to load filament to toolhead.
+- Added raw_load_state, this always returns the current state of the load sensor. Updated some load_states to raw_load_state.
+- Rearranged timeout in prep_callback to help stop message popping up that lane cannot load because printer is moving or homing.
+- Added logic to ViViD units to try multiple times to load filament to load sensor.
+### Fixed
+- Fixed issue where lane would not stop retracting back when running PREP and AFC tries to fix lane. Updated moves to be homing if homing is enabled.
+- Fixed issue where internal state was not being set correctly when fixing a lane during PREP.
+
 ## [2026-02-27]
 ### Added
 - New `lower_extruder_temp_on_change` config option in `AFC.cfg`. When set to `False`, AFC will not lower the extruder temperature during a filament change as long as the current temperature is already sufficient for the target material (within 5°C). Defaults to `True` to preserve existing behaviour.